Server improvment

+++ client/src/main/java/Network/ClientPacketHandler.java
@@ -14,11 +14,11 @@ import simulation.ScheduledActions.ScheduledAction;
 import simulation.Simulation.Snapshot;
 
 public class ClientPacketHandler extends SimpleChannelInboundHandler<Packet> {
-    private int code;
+    private long code;
     private Screen screen;
     private Simulation simulation;
 
-    public ClientPacketHandler(int code, Screen screen) {
+    public ClientPacketHandler(long code, Screen screen) {
         this.code = code;
         this.screen = screen;
     }
